#33 wip move classnames to shared lib
