+++ src/components/Chip/Chip.tsx
@@ -1,5 +1,5 @@
 import { type FunctionComponent } from 'react';
-import { classNames } from '../../utils';
+import { classNames } from '../../lib/classNames';
 import type { ChipProps } from './Chip.types';
 import styles from './Chip.module.css';
