+++ src/features/currency/components/DataTable/DataTable.tsx
@@ -4,7 +4,7 @@ import {
   type FunctionComponent,
 } from 'react';
 import { Bulb } from '../../../../components/Bulb';
-import { classNames } from '../../../../utils';
+import { classNames } from '../../../../lib/classNames';
 import { getSeriesColor } from '../../utils';
 import { XSizingRow } from './components/XSizingRow';
 import { YSpacingRow } from './components/YSpacingRow';
